Move loading of service configuration into `service::config` module.
